requirements.txt
Flask == 1.0.2 
requests == 2.23.0
Flask_cors == 3.0.9
line-bot-sdk == 1.17.0 
gunicorn == 20.0.4